requirements.txt
asgiref==3.4.1
cached-property==1.5.2
Django==3.2.6
django-url-filter==0.3.15
djangorestframework==3.12.4
enum-compat==0.0.3
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
